--- components/bentoGrid/bentoGrid.module.css
@@ -1,37 +1,0 @@
-.bentoGrid {
-  display: grid;
-  width: 100%;
-  gap: 16px;
-  margin-top: var(--gap-md);
-  grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(auto-fit, 1fr);
-}
-
-.bentoGridColumn {
-  display: flex;
-  height: fit-content;
-  flex-direction: column;
-  overflow: hidden;
-  grid-column: span 2;
-  grid-row: span 1;
-  gap: 16px;
-}
-
-.bentoGridImage {
-  object-fit: cover;
-  width: 100%;
-  height: 100%;
-  border-radius: 20px;
-}
-
-@media (max-width: 1024px) {
-  .bentoGrid {
-    grid-template-columns: repeat(4, 1fr);
-  }
-}
-
-@media (max-width: 640px) {
-  .bentoGrid {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
